backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
PyJWT
passlib[bcrypt]
python-dotenv
numpy
openai